requirements.txt
Django==4.0.3
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.4.0
